03_Django/02_django_CRUD/02-3_django_static_media/requirements.txt
astroid==2.2.5
attrs==19.1.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
bs4==0.0.1
certifi==2019.6.16
chardet==3.0.4
Click==7.0
colorama==0.4.1
decorator==4.4.0
defusedxml==0.6.0
Django==2.2.5
django-appconf==1.0.3
django-extensions==2.2.1
django-imagekit==4.0.2
entrypoints==0.3
Faker==2.0.2
Flask==1.1.1
idna==2.8
ipykernel==5.1.1
ipython==7.6.1
ipython-genutils==0.2.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.14.1
Jinja2==2.10.1
Js2Py==0.66
json5==0.8.5
jsonschema==3.0.1
jupyter-client==5.3.1
jupyter-contrib-core==0.3.3
jupyter-contrib-nbextensions==0.5.1
jupyter-core==4.5.0
jupyter-highlight-selected-word==0.2.0
jupyter-latex-envs==1.4.6
jupyter-nbextensions-configurator==0.4.1
jupyterlab==1.0.2
jupyterlab-server==1.0.0
lazy-object-proxy==1.4.1
lxml==4.3.4
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.5.0
nbformat==4.4.0
node==0.9.24
nodejs==0.1.1
notebook==5.7.8
odict==1.7.0
optional-django==0.1.0
pandocfilters==1.4.2
parso==0.5.1
pickleshare==0.7.5
pilkit==2.0
Pillow==6.2.1
plumber==1.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
Pygments==2.4.2
pyjsparser==2.7.1
pylint==2.3.1
pyrsistent==0.15.3
python-dateutil==2.8.0
python-decouple==3.1
pytz==2019.2
pywinpty==0.5.5
PyYAML==5.1.1
pyzmq==18.0.2
requests==2.22.0
Send2Trash==1.5.0
six==1.12.0
soupsieve==1.9.2
sqlparse==0.3.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.3
tornado==6.0.3
traitlets==4.3.2
typed-ast==1.4.0
tzlocal==2.0.0
urllib3==1.25.3
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.4
wrapt==1.11.2
zope.component==4.5
zope.deferredimport==4.3.1
zope.deprecation==4.4.0
zope.event==4.4
zope.hookable==4.2.0
zope.interface==4.6.0
zope.lifecycleevent==4.3
zope.proxy==4.3.2
